+++ .lintstagedrc.mjs
@@ -1,6 +1,6 @@
 export default {
     '*.*': filenames => [`case-police  --fix ${filenames.join(' ')}`],
-    '*.{json,mjs,ts}?(x)': filenames => [
+    '*.{json,mjs,ts,vue}': filenames => [
         `eslint ${filenames.join(' ')}`,
         `vitest ${filenames.join(' ')} --passWithNoTests`,
     ],
