+++ .tidierrc
@@ -1,6 +1,7 @@
 {
     "ignore": [
         "**/build",
+        "**/workflow/export",
         "node_modules"
     ],
     "files": {
