+++ nuxt.config.ts
@@ -7,5 +7,15 @@ export default defineNuxtConfig({
         '@pinia/nuxt',
         '@unocss/nuxt',
     ],
-    unocss: { uno: true },
+    unocss: {
+        theme: {
+            breakpoints: {
+                '3xl': '1980px',
+                '4xl': '2560px',
+                '5xl': '3200px',
+            },
+        },
+        uno: true,
+    },
+    vite: { build: { sourcemap: true } },
 });
